fixed serverside auth middleware
